Revert "Revert "Upgrade spdlog""
This reverts commit 5da55c23b3d4a8aca13ea974709db3a4f99d4d5b.

+++ CesiumRasterOverlays/src/IonRasterOverlay.cpp
@@ -127,7 +127,7 @@ IonRasterOverlay::createTileProvider(
                   fmt::format(
                       "Error while parsing Cesium ion raster overlay response, "
                       "error code {} at byte offset {}",
-                      response.GetParseError(),
+                      static_cast<int32_t>(response.GetParseError()),
                       response.GetErrorOffset())});
             }
